added gofr-errors

+++ gofr-test/Internal/service/movieservice/service.go
@@ -1,8 +1,8 @@
 package movieservice
 
 import (
+	"developer.zopsmart.com/go/gofr/pkg/errors"
 	"developer.zopsmart.com/go/gofr/pkg/gofr"
-	"errors"
 	"golangprog/gofr-test/Internal/models"
 	stores "golangprog/gofr-test/Internal/store"
 )
@@ -17,7 +17,7 @@ func New(store stores.Movie) *Service {
 
 func (se *Service) GetByID(ctx *gofr.Context, id int) (*models.Movie, error) {
 	if id <= 0 {
-		return nil, errors.New("error invalid id")
+		return nil, errors.InvalidParam{Param: []string{"id"}}
 	}
 
 	movieObj, err := se.store.GetByID(ctx, id)
@@ -40,10 +40,13 @@ func (se *Service) GetAll(ctx *gofr.Context) ([]*models.Movie, error) {
 }
 
 func (se *Service) Delete(ctx *gofr.Context, id int) error {
+	if id <= 0 {
+		return errors.InvalidParam{Param: []string{"id"}}
+	}
 	err := se.store.Delete(ctx, id)
 
 	if err != nil {
-		return errors.New("error invalid id")
+		return err
 	}
 
 	return nil
@@ -51,11 +54,11 @@ func (se *Service) Delete(ctx *gofr.Context, id int) error {
 
 func (se *Service) Create(ctx *gofr.Context, movieObj *models.Movie) (*models.Movie, error) {
 	if movieObj.Name == "" {
-		return nil, errors.New("error invalid name")
+		return nil, errors.InvalidParam{}
 	}
 
 	if movieObj.Plot == "" {
-		return nil, errors.New("error invalid plot")
+		return nil, errors.InvalidParam{}
 	}
 
 	movieObjs, err := se.store.Create(ctx, movieObj)
@@ -69,15 +72,15 @@ func (se *Service) Create(ctx *gofr.Context, movieObj *models.Movie) (*models.Mo
 
 func (se *Service) Update(ctx *gofr.Context, movieObj *models.Movie) (*models.Movie, error) {
 	if movieObj.ID < 0 {
-		return nil, errors.New("error invalid id")
+		return nil, errors.InvalidParam{}
 	}
 
 	if movieObj.Name == "" {
-		return nil, errors.New("error invalid name")
+		return nil, errors.InvalidParam{}
 	}
 
 	if movieObj.Plot == "" {
-		return nil, errors.New("error invalid plot")
+		return nil, errors.InvalidParam{}
 	}
 
 	movieObjs, err := se.store.Update(ctx, movieObj)
